+++ src/video_server/process.clj
@@ -85,19 +85,41 @@
   (let [containers (filter web-playback? (:containers video))]
     (some #(and (= (file-type (:path %)) fmt) (= size (-> % :width video-size))) containers)))
 
+(defn- smallest-size
+  [video fmt]
+  (->> (:containers video)
+       (filter #(= fmt (:filetype %)))
+       (map (comp #(Integer/parseInt %) name video-size :width))
+       sort first str keyword))
+
 (defn- next-encode-size
   "Returns the next size for the default encoding."
   [video fmt size]
-  (if (has-fmt-size? video fmt size) ({:2160 :1080 :1080 :720} size) size))
+  (if (has-fmt-size? video fmt size) ({:2160 :1080 :1080 :720 :720 :480} size) size))
 
 (defn should-encode-video?
   "Returns whether the video should be encoded for downloading or
   casting."
   ([video]
    (not (and (can-download? video) (can-cast? video))))
+  ([video [fmt size]]
+    (should-encode-video? video fmt size))
   ([video fmt size]
    (and (not (has-fmt-size? video fmt size))
         (can-source? (-> video :containers (container-to-encode size) :width video-size) size))))
+
+(defn- next-encoding
+  "Returns the next [fmt size] for the default encoding, or nil."
+  [video fmt size apple? download min-download?]
+  (when (and fmt size)
+    (->> [(when-not (has-fmt-size? video fmt size) [fmt size false])
+          (when (and apple? (= fmt :mkv)) [:m4v size false])
+          (when (and download (not (can-download? video download)))
+            (if-let [next-size (next-encode-size video fmt (smallest-size video fmt))]
+              [fmt next-size min-download?]))]
+         (remove nil?)
+         (filter (partial should-encode-video? video))
+         first)))
 
 (defn should-encode-subtitles?
   "Returns whether the subtitles need to be transcoded for casting."
@@ -107,28 +129,28 @@
 
 (defn- encode
   "Creates an encoding spec and queues it for processing."
-  [folder video fmt size]
-  (when-let [spec (video-encode-spec folder video fmt size)]
+  [folder video fmt size minimize?]
+  (when-let [spec (video-encode-spec folder video fmt size minimize?)]
     (process-encode spec)))
 
 (defn process-video
   "Conditionally encodes the video and/or subtitles."
-  [folder video fmt size options]
+  [folder video fmt size apple? download min-download? containers]
   (log/debug "processing video" (str video))
   (when (should-encode-subtitles? video)
     (encode-subtitles folder video))
-  (let [size (next-encode-size video fmt size)]
-    (if (and size (should-encode-video? video fmt size))
-      (encode folder video fmt (encode-size video size))
+  (let [[fmt size minimize?] (next-encoding video fmt size apple? download min-download?)]
+    (if (and size fmt)
+      (encode folder video fmt (encode-size video size) minimize?)
       (when-not *fake-encode*
-        (loop [options options]
-          (when (seq options)
-            (let [{fmt :format size :size} (first options)
+        (loop [containers containers]
+          (when (seq containers)
+            (let [{fmt :format size :size} (first containers)
                   size (encode-size video size)]
               (log/trace "checking" fmt size "for" (:title video))
               (if (should-encode-video? video fmt size)
-                (encode folder video fmt size)
-                (recur (rest options))))))))))
+                (encode folder video fmt size false)
+                (recur (rest containers))))))))))
 
 (defn- start-fetching
   "Processes requests in the fetch channel using real threads."
@@ -169,7 +191,7 @@
 
 (defn start-processing
   "Processes enqueued files."
-  [encode? fetch? fmt size]
+  [fmt size {encode? :encode fetch? :fetch apple? :apple download :download min-download? :min-download}]
   (log/debug "starting file processing")
   (start-fetching movie-chan fetch-threads fetch?)
   (start-fetching series-chan 1 fetch?)
@@ -186,7 +208,9 @@
              (when (get options :encode encode?)
                (process-video folder video
                               (get options :format fmt) (get options :size size)
-                              (-> folder :options :encoder :containers)))
+                              (get options :apple apple?)
+                              (get options :download download) (get options :min-download min-download?)
+                              (-> options :encoder :containers)))
              (catch Exception e (log/error e "error processing video" (str video))))))
     (recur)))